- Now that we got time_t to work nicely on openbsd with snprintf's conversion specifiers,   we ran into a similiar issue on Raspbian/ARMv7's time_t which is of signed 32 bit and   doesn't cope at all with %j. Instead of doing tricks, get rid of time_t everywhere and   forever and use uintmax_t instead which has at least a 'standardized' conversion specifier   associated with it.
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/branches/8.2.x@7329 82007160-df01-0410-b94d-b575c5fd34c7

+++ tools/mkpasswd.c
@@ -315,7 +315,7 @@ generate_random_salt(char *salt, unsigned int length)
 
   buf = calloc(1, length);
 
-  if (read(fd, buf, length) != length)
+  if (read(fd, buf, length) != (ssize_t)length)
   {
     close(fd);
     free(buf);
